Inconsistent behaviour in Exceptions when using Exception.exception/1 and Exception.message/1 callbacks
This issues is closely related to https://github.com/elixir-lang/elixir/issues/3657

It is recommended by that an exception can be defined with either `Exception.exception/1` or `Exception.message/1` call back,
While they raise with no problem they behave accordingly, when there is a problem building the exception, they will behave different, and the users has no way to know which one to except.

```elixir
iex(1)> raise CompileError
** (CompileError) got FunctionClauseError with message `no function clause matching in IO.chardata_to_string/1` while retrieving Exception.message/1 for %CompileError{description: "compile error", file: nil, line: nil}

iex(1)> raise Code.LoadError
** (KeyError) key :file not found in: []
    (elixir) lib/keyword.ex:235: Keyword.fetch!/2
    (elixir) lib/exception.ex:659: Code.LoadError.exception/1
```

both raises an exception, because the required  `:file` field was not provided, but they behave differently because `message/1` catches the error, and returns the original exception being called (`CompileError`),
and `exception/1` doesn't do any handling and pass the new exception directly (`KeyError`).

I think we we should either make `exception/1` try/rescue as in `message/1`, or reraise in `message/1` with the new exception that was just thrown.

I have created an test for this
https://github.com/eksperimental/elixir_fail/blob/master/test/exception_vs_message.exs

* reported issue: https://github.com/elixir-lang/elixir/issues/3663